Revert "Move IndexMetadata, NodesResponse, SearchShardsResponse, Shard, ElasticsearchNode into Elasticsearch client directory"
This reverts commit eba938f593336df9c9488714ad2dc39224085ba7.
