+++ presto-elasticsearch/src/main/java/com/facebook/presto/elasticsearch/ElasticsearchClient.java
@@ -17,11 +17,6 @@ import com.facebook.airlift.json.JsonCodec;
 import com.facebook.airlift.json.ObjectMapperProvider;
 import com.facebook.airlift.log.Logger;
 import com.facebook.airlift.security.pem.PemReader;
-import com.facebook.presto.elasticsearch.client.ElasticsearchNode;
-import com.facebook.presto.elasticsearch.client.IndexMetadata;
-import com.facebook.presto.elasticsearch.client.NodesResponse;
-import com.facebook.presto.elasticsearch.client.SearchShardsResponse;
-import com.facebook.presto.elasticsearch.client.Shard;
 import com.facebook.presto.spi.PrestoException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -104,7 +99,7 @@ public class ElasticsearchClient
     private final int scrollSize;
     private final Duration scrollTimeout;
 
-    private final AtomicReference<Set<ElasticsearchNode>> nodes = new AtomicReference<>(ImmutableSet.of());
+    private final AtomicReference<Set<Node>> nodes = new AtomicReference<>(ImmutableSet.of());
     private final ScheduledExecutorService executor = newSingleThreadScheduledExecutor(daemonThreadsNamed("NodeRefresher"));
     private final AtomicBoolean started = new AtomicBoolean();
     private final Duration refreshInterval;
@@ -143,10 +138,10 @@ public class ElasticsearchClient
     {
         // discover other nodes in the cluster and add them to the client
         try {
-            Set<ElasticsearchNode> nodes = fetchNodes();
+            Set<Node> nodes = fetchNodes();
 
             HttpHost[] hosts = nodes.stream()
-                    .map(ElasticsearchNode::getAddress)
+                    .map(Node::getAddress)
                     .map(address -> HttpHost.create(format("%s://%s", tlsEnabled ? "https" : "http", address)))
                     .toArray(HttpHost[]::new);
 
@@ -298,36 +293,36 @@ public class ElasticsearchClient
         }
     }
 
-    private Set<ElasticsearchNode> fetchNodes()
+    private Set<Node> fetchNodes()
     {
         NodesResponse nodesResponse = doRequest("_nodes/http", NODES_RESPONSE_CODEC::fromJson);
 
-        ImmutableSet.Builder<ElasticsearchNode> result = ImmutableSet.builder();
+        ImmutableSet.Builder<Node> result = ImmutableSet.builder();
         for (Map.Entry<String, NodesResponse.Node> entry : nodesResponse.getNodes().entrySet()) {
             String nodeId = entry.getKey();
             NodesResponse.Node node = entry.getValue();
 
             if (node.getRoles().contains("data")) {
-                result.add(new ElasticsearchNode(nodeId, node.getHttp().getAddress()));
+                result.add(new Node(nodeId, node.getHttp().getAddress()));
             }
         }
         return result.build();
     }
 
-    public Set<ElasticsearchNode> getNodes()
+    public Set<Node> getNodes()
     {
         return nodes.get();
     }
 
     public List<Shard> getSearchShards(String index)
     {
-        Map<String, ElasticsearchNode> nodeById = getNodes().stream()
-                .collect(toImmutableMap(ElasticsearchNode::getId, Function.identity()));
+        Map<String, Node> nodeById = getNodes().stream()
+                .collect(toImmutableMap(Node::getId, Function.identity()));
 
         SearchShardsResponse shardsResponse = doRequest(format("%s/_search_shards", index), SEARCH_SHARDS_RESPONSE_CODEC::fromJson);
 
         ImmutableList.Builder<Shard> shards = ImmutableList.builder();
-        List<ElasticsearchNode> nodes = ImmutableList.copyOf(nodeById.values());
+        List<Node> nodes = ImmutableList.copyOf(nodeById.values());
 
         for (List<SearchShardsResponse.Shard> shardGroup : shardsResponse.getShardGroups()) {
             Stream<SearchShardsResponse.Shard> preferred = shardGroup.stream()
@@ -338,7 +333,7 @@ public class ElasticsearchClient
                     .findFirst();
 
             SearchShardsResponse.Shard chosen;
-            ElasticsearchNode node;
+            Node node;
             if (candidate.isPresent()) {
                 chosen = candidate.get();
                 node = nodeById.get(chosen.getNode());
