+++ presto-elasticsearch/src/main/java/com/facebook/presto/elasticsearch/NodesResponse.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.elasticsearch.client;
+package com.facebook.presto.elasticsearch;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
